move persistence apis into its sep class
